requirements.txt
requests>=2.31.0
PyPDF2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
